// src/common/enums/certificationType.enum.ts
export enum CertificationType {
  DRIVING_LICENSE = 'driving_license',
  PROFESSIONAL_LICENSE = 'professional_license',
  DANGEROUS_GOODS = 'dangerous_goods',
  MEDICAL_EXAM = 'medical_exam',
  HAZMAT = 'hazmat',
  CRANE_OPERATOR = 'crane_operator',
  DEFENSIVE_DRIVING = 'defensive_driving',
  FIRST_AID = 'first_aid',
  OTHER = 'other',
}